+++ docs/pages/map-providers/_meta.js
@@ -3,5 +3,6 @@ export default {
   geobase: "Geobase",
   esri: "ESRI",
   tms: "TMS",
+  wms: "WMS",
   "serving-raster-tiles": "Serving Raster Tiles",
 };